+++ packages/panvala-utils/voting/index.ts
@@ -82,6 +82,56 @@ function encodeBallot(resources: string[], firstChoices: string[], secondChoices
   return encoded;
 }
 
+const BN = (small: utils.BigNumberish) => utils.bigNumberify(small);
+
+/**
+ * Calculate the next slate submission deadline as halfway between now and the start of the
+ * commit period.
+ */
+function slateSubmissionDeadline(votingOpenDate: number, lastStaked: number) {
+  // prettier-ignore
+  const extraTime = BN(votingOpenDate).sub(BN(lastStaked)).div('2');
+  return BN(lastStaked)
+    .add(extraTime)
+    .toNumber();
+}
+
+interface IBallotDates {
+  startDate: number;
+  votingOpenDate: number;
+  votingCloseDate: number;
+  finalityDate: number;
+  initialSlateSubmissionDeadline: number;
+  slateSubmissionDeadline: {
+    [key: string]: number;
+  };
+  epochNumber: number;
+}
+
+function ballotDates(startDate: number = 1549040400): IBallotDates {
+  const oneWeekSeconds = 604800;
+  const epochStartDate = utils.bigNumberify(startDate).toNumber();
+  const week11EndDate: number = epochStartDate + oneWeekSeconds * 11; // 1555689600
+  const week12EndDate: number = week11EndDate + oneWeekSeconds;
+  const week13EndDate: number = week12EndDate + oneWeekSeconds;
+  const initialSlateSubmissionDeadline = slateSubmissionDeadline(week11EndDate, startDate);
+
+  return {
+    startDate: epochStartDate,
+    votingOpenDate: week11EndDate,
+    votingCloseDate: week12EndDate,
+    finalityDate: week13EndDate,
+    initialSlateSubmissionDeadline,
+    // TODO: use the resource (addresses) instead of GRANT/GOVERNANCE
+    slateSubmissionDeadline: {
+      GRANT: 0,
+      GOVERNANCE: 0,
+    },
+    epochNumber: 0,
+  };
+}
+
+
 module.exports = {
   generateCommitHash,
   randomSalt,
@@ -89,4 +139,6 @@ module.exports = {
   encodeBallot,
   SlateCategories,
   ContestStatus,
+  slateSubmissionDeadline,
+  ballotDates,
 };
